rename the unit test file to *.ts
